refactor: harden openclaw shadow plugin
